+++ examples/cfd/darcy_transolver/train_transolver_darcy.py
@@ -70,6 +70,7 @@ def darcy_trainer(cfg: DictConfig) -> None:
         structured_shape=[cfg.data.resolution, cfg.data.resolution],
         use_te=cfg.model.use_te,
         time_input=cfg.model.time_input,
+        activation_checkpointing=cfg.model.get("activation_checkpointing", False),
     ).to(dist.device)
 
     loss_fun = TestLoss(size_average=False)